Make `getGeneratedAnno()` more strongly typed.
Also renamed some classes.
