+++ at-paninij-core/at-paninij-proc/src/main/java/org/paninij/proc/factory/CapsuleInterfaceFactory.java
@@ -16,19 +16,20 @@
  *
  * Contributor(s): Dalton Mills
  */
-package org.paninij.proc;
+package org.paninij.proc.factory;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import org.paninij.proc.PaniniProcessor;
 import org.paninij.proc.model.Procedure;
 import org.paninij.proc.model.Variable;
 import org.paninij.proc.util.MessageShape;
 import org.paninij.proc.util.Source;
 
-public class CapsuleInterfaceFactory extends CapsuleArtifactFactory
+public class CapsuleInterfaceFactory extends AbstractCapsuleFactory
 {
     @Override
     protected String getQualifiedName()
@@ -55,7 +56,7 @@ public class CapsuleInterfaceFactory extends CapsuleArtifactFactory
 
         src = Source.format(src,
                 this.capsule.getPackage(),
-                PaniniProcessor.getGeneratedAnno("CapsuleInterfaceFactory"),
+                PaniniProcessor.getGeneratedAnno(CapsuleInterfaceFactory.class),
                 this.capsule.getSimpleName(),
                 this.generateInterfaces(),
                 this.generateImportDecl());
