+++ at-paninij-core/at-paninij-proc/src/main/java/org/paninij/proc/factory/CapsuleTestFactory.java
@@ -16,18 +16,19 @@
  *
  * Contributor(s): Dalton Mills, David Johnston
  */
-package org.paninij.proc;
+package org.paninij.proc.factory;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import org.paninij.proc.PaniniProcessor;
 import org.paninij.proc.model.Procedure;
 import org.paninij.proc.util.MessageShape;
 import org.paninij.proc.util.Source;
 
-public class CapsuleTestFactory extends CapsuleArtifactFactory
+public class CapsuleTestFactory extends AbstractCapsuleFactory
 {
     public static final String CAPSULE_TEST_SUFFIX = "$Test";
 
@@ -51,7 +52,7 @@ public class CapsuleTestFactory extends CapsuleArtifactFactory
 
         src = Source.format(src,
                 this.capsule.getPackage(),
-                PaniniProcessor.getGeneratedAnno("CapsuleTestFactory"),
+                PaniniProcessor.getGeneratedAnno(CapsuleTestFactory.class),
                 this.generateClassName());
         src = Source.formatAligned(src, generateImports());
         src = Source.formatAligned(src, generateTests());
