+++ at-paninij-core/at-paninij-proc/src/main/java/org/paninij/proc/factory/CapsuleThreadFactory.java
@@ -16,14 +16,14 @@
  *
  * Contributor(s): Dalton Mills, David Johnston, Trey Erenberger
  */
-package org.paninij.proc;
+package org.paninij.proc.factory;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import org.paninij.proc.model.Behavior;
+import org.paninij.proc.PaniniProcessor;
 import org.paninij.proc.model.Procedure;
 import org.paninij.proc.model.Type;
 import org.paninij.proc.model.Variable;
@@ -31,14 +31,14 @@ import org.paninij.proc.util.MessageShape;
 import org.paninij.proc.util.PaniniModel;
 import org.paninij.proc.util.Source;
 
-public class CapsuleTaskFactory extends CapsuleProfileFactory
+public class CapsuleThreadFactory extends CapsuleProfileFactory
 {
-    public static final String CAPSULE_PROFILE_TASK_SUFFIX = "$Task";
+    public static final String CAPSULE_PROFILE_THREAD_SUFFIX = "$Thread";
 
     @Override
     protected String getQualifiedName()
     {
-        return this.capsule.getQualifiedName() + CAPSULE_PROFILE_TASK_SUFFIX;
+        return this.capsule.getQualifiedName() + CAPSULE_PROFILE_THREAD_SUFFIX;
     }
 
     @Override
@@ -51,14 +51,15 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
                 "",
                 "#1",
                 "@SuppressWarnings(\"unused\")",  // To suppress unused import warnings.
-                "public class #2 extends Capsule$Task implements #3",
+                "@CapsuleThread",
+                "public class #2 extends Capsule$Thread implements #3",
                 "{",
                 "    ##",
                 "}");
 
         src = Source.format(src,
                 this.capsule.getPackage(),
-                PaniniProcessor.getGeneratedAnno("CapsuleTaskFactory"),
+                PaniniProcessor.getGeneratedAnno(CapsuleThreadFactory.class),
                 this.generateClassName(),
                 this.capsule.getSimpleName());
 
@@ -71,7 +72,7 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
     @Override
     protected String generateClassName()
     {
-        return this.capsule.getSimpleName() + CAPSULE_PROFILE_TASK_SUFFIX;
+        return this.capsule.getSimpleName() + CAPSULE_PROFILE_THREAD_SUFFIX;
     }
 
     private List<String> generateImports()
@@ -87,11 +88,13 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
         
         imports.add("javax.annotation.Generated");
         imports.add("java.util.concurrent.Future");
-        imports.add("org.paninij.runtime.Capsule$Task");
+        imports.add("org.paninij.lang.CapsuleThread");
+        imports.add("org.paninij.runtime.Capsule$Thread");
         imports.add("org.paninij.runtime.Panini$Capsule");
         imports.add("org.paninij.runtime.Panini$Message");
         imports.add("org.paninij.runtime.Panini$Future");
         imports.add("org.paninij.runtime.Panini$System");
+        imports.add("org.paninij.runtime.check.DynamicOwnershipTransfer");
         imports.add(this.capsule.getQualifiedName());
 
         List<String> prefixedImports = new ArrayList<String>();
@@ -126,33 +129,6 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
         return decls;
     }
 
-    @Override
-    protected List<String> generateProcedure(Procedure procedure) {
-        MessageShape shape = new MessageShape(procedure);
-        String doBlock = shape.behavior == Behavior.BLOCKED_FUTURE || shape.behavior == Behavior.BLOCKED_PREMADE ? "panini$emptyQueue();" : "";
-        String encoding = PaniniModel.isPaniniCustom(shape.returnType.getMirror()) ? shape.returnType.raw() : shape.encoded;
-        
-        List<String> source = Source.lines(
-                "@Override",
-                "#0",
-                "{",
-                "    #1 panini$message = null;",
-                "    panini$message = new #1(#2);",
-                "    #3;",
-                "    panini$push(panini$message);",
-                "    #4",
-                "    #5",
-                "}",
-                "");
-        return Source.formatAll(source,
-                this.generateProcedureDecl(shape),
-                encoding,
-                this.generateProcedureArguments(shape),
-                this.generateAssertSafeInvocationTransfer(),
-                doBlock,
-                this.generateProcedureReturn(shape));
-    }
-
     private List<String> generateProcedures()
     {
         ArrayList<String> src = new ArrayList<String>();
@@ -180,13 +156,13 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
                         lines,
                         local.getIdentifier(),
                         local.getEncapsulatedType(),
-                        CAPSULE_PROFILE_TASK_SUFFIX));
+                        CAPSULE_PROFILE_THREAD_SUFFIX));
             } else {
                 source.add(Source.format(
                         "panini$encapsulated.#0 = new #1#2();",
                         local.getIdentifier(),
                         local.raw(),
-                        CAPSULE_PROFILE_TASK_SUFFIX));
+                        CAPSULE_PROFILE_THREAD_SUFFIX));
             }
         }
 
@@ -240,22 +216,23 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
         if (this.capsule.isActive()) {
             return Source.lines(
                     "@Override",
-                    "public final boolean run() {",
+                    "public void run() {",
+                    "    Panini$System.self.set(this);",
                     "    try {",
-                    "        panini$capsuleInit();",
+                    "        panini$checkRequiredFields();",
+                    "        panini$initLocals();",
+                    "        panini$initState();",
                     "        panini$encapsulated.run();",
                     "    } catch (Throwable thrown) {",
-                    "        thrown.printStackTrace(System.out);",
                     "        panini$errors.add(thrown);",
                     "    } finally {",
                     "        panini$onTerminate();",
+                    "        try {",
+                    "           Panini$System.threads.countDown();",
+                    "        } catch (InterruptedException e) {",
+                    "            e.printStackTrace();",
+                    "        }",
                     "    }",
-                    "    try {",
-                    "       Panini$System.threads.countDown();",
-                    "    } catch (InterruptedException e) {",
-                    "        e.printStackTrace();",
-                    "    }",
-                    "    return true;",
                     "}",
                     "");
         }
@@ -263,14 +240,26 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
         List<String> src = Source.lines(
                 "@Override",
                 "@SuppressWarnings(\"unchecked\")",
-                "public final boolean run() {",
+                "public void run() {",
+                "    Panini$System.self.set(this);",
                 "    try {",
-                "        Panini$Message msg = panini$nextMessage();",
-                "        ##",
+                "        panini$checkRequiredFields();",
+                "        panini$initLocals();",
+                "        panini$initState();",
+                "",
+                "        boolean terminated = false;",
+                "        while (!terminated) {",
+                "            Panini$Message msg = panini$nextMessage();",
+                "            ##",
+                "        }",
                 "    } catch (Throwable thrown) {",
                 "        panini$errors.add(thrown);",
                 "    }",
-                "    return false;",
+                "    try {",
+                "       Panini$System.threads.countDown();",
+                "    } catch (InterruptedException e) {",
+                "        e.printStackTrace();",
+                "    }",
                 "}",
                 "");
 
@@ -294,7 +283,8 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
                 "    break;",
                 "case PANINI$TERMINATE:",
                 "    panini$onTerminate();",
-                "    return true;",
+                "    terminated = true;",
+                "    break;",
                 "}"));
         return lines;
     }
@@ -376,10 +366,11 @@ public class CapsuleTaskFactory extends CapsuleProfileFactory
 
     private String generateAssertSafeResultTransfer()
     {
-        /*
+        // TODO: Clean this up!
+        /**
         return Source.format(
-                "assert Panini$Ownership.#0.isSafeTransfer(#1, #2) : #3",
-                PaniniProcessor.ownershipCheckMethod,
+                "assert DynamicOwnershipTransfer.#0.isSafeTransfer(#1, #2) : #3",
+                PaniniProcessor.dynamicOwnershipTransferKind,
                 "result",
                 "panini$getAllState()",
                 "\"Procedure return attempted unsafe ownership transfer.\"");
